Break up objectives into smaller files
