+++ desc/objectives/utils.py
@@ -7,7 +7,7 @@ from .linear_objectives import (
     FixedIota,
     FixedPsi,
 )
-from .nonlinear_objectives import RadialForceBalance, HelicalForceBalance, Energy
+from ._equilibrium import RadialForceBalance, HelicalForceBalance, Energy
 
 
 def get_fixed_boundary_constraints():
